Added Backend color sensor files back into ColorSensor folder.

+++ ROBOT_2013/Libraries/Color Sensor/BackEndColorSensor.h
@@ -1,11 +1,11 @@
 #ifndef BACKENDCOLORSENSOR_H
 #define BACKENDCOLORSENSOR_H
-
+#include <Arduino.h>
 class BackEndColorSensor
 {
     public:
         BackEndColorSensor(); //Nescessary color sensor setup called in constructor
-        int detectColor(int color);
+        float detectColor(int color);
     private:
         void TCS3200setup();
         float colorRead(int taosOutPin, int color, boolean LEDstate);
